PRSD-406: Move RadioButtonDataModel to models.formModels
