+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/models/formModels/RadioButtonDataModel.kt
@@ -1,4 +1,4 @@
-package uk.gov.communities.prsdb.webapp.models.dataModels
+package uk.gov.communities.prsdb.webapp.models.formModels
 
 data class RadioButtonDataModel<T>(
     val value: T,
